refactor!: rename observer to ValidateModel
This is a better name, because it gets the point across that the class
does something as opposed to being an entity that maintains state. It
validates the given model while the model is being saved.

BREAKING CHANGE: class name referenced by client code is changed. The
previous name will raise a fatal error.

Resolves #12