refactor: rename app directory to src
This is not an application; therefore, src makes way more sense than
app. I used app previously because artisan was expecting it and I could
not see a way to configure that. The package will then better match the
standard PHP package skeleton.

To make things work, I ran commands:

    composer dump-autoload
    artisan config:clear
    artisan cache:clear